requirements.txt
FLASK
flask_cors
pillow
tensorflow
numpy
opencv-python
gunicorn